+++ runtime/layers/fpm/bootstrap
@@ -1,7 +1,7 @@
 #!/opt/bin/php
 <?php declare(strict_types=1);
 
-use Bref\Handler\FpmHandler;
+use Bref\Event\Http\FpmHandler;
 use Bref\Runtime\LambdaRuntime;
 
 ini_set('display_errors', '1');
