+++ src/Event/Handler.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Bref\Handler;
+namespace Bref\Event;
 
 use Bref\Context\Context;
